pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stageflow"
version = "0.1.0"
description = "StageFlow: Pipeline framework for stages"
authors = [
    { name="lethargy" }
]
dependencies = [
    "pyyaml",
    "jsonschema>=4.0.0"
]
requires-python = ">=3.10"

[tool.setuptools.packages.find]
where = ["."]
include = ["stageflow*"]

[tool.setuptools.package-data]
"stageflow" = ["py.typed"]
